Standardizing navbar, bug fixes.
Making the navbar the same (blue, navbar2.php) for all users who arent
logged in, on the welcome, login, and create user page. Making
navbar.php (not the 2nd one) black, and on the rest of the pages when
the user is logged in. User.php was missing some semicolons.